+++ tests/db/test_db_schema.py
@@ -32,7 +32,7 @@ import inmanta.db.versions
 from data.db import versions
 from inmanta import data
 from inmanta.data import CORE_SCHEMA_NAME, schema
-from inmanta.data.schema import TableNotFound, Version
+from inmanta.data.schema import InvalidSchemaVersion, TableNotFound, Version, create_schemamanager
 from utils import log_contains
 
 
@@ -357,3 +357,21 @@ async def test_dbschema_get_dct_filter_disabled():
         assert isinstance(version.function, types.FunctionType)
         assert version.function.__name__ == "update"
         assert inspect.getfullargspec(version.function)[0] == ["connection"]
+
+
+@pytest.mark.asyncio
+async def test_dbschema_update_db_downgrade(postgresql_client):
+    schema_name = "test_dbschema_update_db_downgrade"
+    SCHEMA_VERSION_TABLE = "schemamanager"
+    await postgresql_client.execute(create_schemamanager)
+
+    db_schema = schema.DBSchema(schema_name, inmanta.db.versions, postgresql_client)
+    update_function_map = await db_schema._get_update_functions()
+    original_version = len(update_function_map) + 1
+    await postgresql_client.execute(
+        f"INSERT INTO {SCHEMA_VERSION_TABLE} (name, current_version) VALUES ($1, $2)", schema_name, original_version
+    )
+    with pytest.raises(InvalidSchemaVersion):
+        await db_schema.ensure_db_schema()
+    current_db_version = await db_schema.get_current_version()
+    assert original_version == current_db_version